New connection
New room added leading to large container room next to cp2. This should
give the blu team an easyer way to get on tho the high fround inside the
room. I will ad a one way door so only the attacking side can use it to
hinder snipers to kill of people flanking cp2. Also it leads to  a
sentry spot for the blu team so the door will block spam.